requirements/test-req.txt
coverage==6.3.2
Flask==2.0.3
Flask-JWT-Extended==4.3.1
flask-restx==0.5.1
marshmallow==3.14.1
psycopg==3.0.9
PyJWT==2.3.0
pytest==7.0.1
pytest-cov==3.0.0
WebTest==3.0.0
